ppomppu/requirements.txt
amqp==2.4.1
awscli==1.16.109
beautifulsoup4==4.7.1
billiard==3.5.0.5
botocore==1.12.99
bs4==0.0.1
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
Django==2.1.7
django-allauth==0.38.0
django-celery-beat==1.4.0
django-cors-headers==2.5.0
django-rest-auth==0.9.3
django-rest-framework==0.1.0
django-timezone-field==3.0
djangorestframework==3.9.1
docutils==0.14
idna==2.8
jmespath==0.9.3
kombu==4.3.0
oauthlib==3.0.1
psycopg2==2.7.7
pyasn1==0.4.5
python-crontab==2.3.6
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==3.13
redis==3.2.0
requests==2.21.0
requests-oauthlib==1.2.0
rsa==3.4.2
s3transfer==0.2.0
six==1.12.0
soupsieve==1.8
urllib3==1.24.1
vine==1.2.0
